+++ go.mod
@@ -4,6 +4,7 @@ go 1.16
 
 require (
 	github.com/google/go-cmp v0.5.6 // indirect
+	github.com/gorilla/mux v1.8.0
 	github.com/prometheus/client_golang v1.11.0
 	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
 )
